+++ src/screen.c
@@ -1,5 +1,5 @@
 #include <gos/screen.h>
 
-int gos_screen_index(gos_screen* screen, int x, int y) {
-  return x + y * screen->width;
+size_t gos_screen_index(gos_screen* screen, int x, int y) {
+  return x + y * ((size_t)screen->width);
 }
